Unduplicate test serializer logic, move to test_rest_framework_dso/serializers.py
